refactor-only; prepare for multiple front-ends, refs #35
