+++ lib/asyncapi1/generator.js
@@ -6,7 +6,7 @@ const _ = require('lodash');
 const randomName = require('project-name-generator');
 const xfs = require('fs.extra');
 const beautifier = require('./beautifier');
-const registerPartial = require('./register-partial');
+const registerPartial = require('../register-partial');
 const parse = require('./parser');
 const minimatch = require("minimatch");
 
@@ -252,7 +252,7 @@ const registerHelpers = config => new Promise((resolve, reject) => {
  */
 const registerPartials = config => new Promise((resolve, reject) => {
   const partials_dir = path.resolve(config.templates, PARTIALS_DIRNAME);
-  
+
   if (!fs.existsSync(partials_dir)) return resolve();
 
   const walker = xfs.walk(partials_dir, {
@@ -298,7 +298,7 @@ const bundleAndApplyDefaults = config => new Promise((resolve, reject) => {
           name: _.kebabCase(_.result(config, 'asyncapi.info.title', random_name))
         },
         target_dir: path.resolve(os.tmpdir(), 'asyncapi-generated'),
-        templates: path.resolve(__dirname, '../templates')
+        templates: path.resolve(__dirname, '../../templates')
       });
 
       config.templates = `${config.templates}/${config.template}`;
